feat: Add astyle

+++ app/src/circular_buffer.c
@@ -9,9 +9,9 @@
 #undef MAX
 #endif
 #define MAX(a,b) \
-   ({ __typeof__ (a) _a = (a); \
-       __typeof__ (b) _b = (b); \
-     _a > _b ? _a : _b; })
+    ({ __typeof__ (a) _a = (a); \
+        __typeof__ (b) _b = (b); \
+        _a > _b ? _a : _b; })
 
 /* GNU GCC MIN(a,b) recomendation for a safe macro
  * https://gcc.gnu.org/onlinedocs/gcc/Typeof.html */
@@ -19,12 +19,13 @@
 #undef MIN
 #endif
 #define MIN(a,b) \
-   ({ __typeof__ (a) _a = (a); \
-       __typeof__ (b) _b = (b); \
-     _a < _b ? _a : _b; })
+    ({ __typeof__ (a) _a = (a); \
+        __typeof__ (b) _b = (b); \
+        _a < _b ? _a : _b; })
 
 /* Space available in circular buffer */
-size_t circ_buff_free_space(circular_buffer_t *buff) {
+size_t circ_buff_free_space(circular_buffer_t *buff)
+{
     if ((buff->tail == buff->head) && (buff->full == true)) {
         return 0;
     }
@@ -37,7 +38,8 @@ size_t circ_buff_free_space(circular_buffer_t *buff) {
 }
 
 /* Current size of circular buffer */
-size_t circ_buff_current_size(circular_buffer_t *buff) {
+size_t circ_buff_current_size(circular_buffer_t *buff)
+{
     if ((buff->tail == buff->head) && (buff->full == true)) {
         return buff->max_size;
     }
@@ -50,12 +52,14 @@ size_t circ_buff_current_size(circular_buffer_t *buff) {
 }
 
 /* Maximum capacity of circular buffer */
-size_t circ_buffer_max_size(circular_buffer_t *buff) {
+size_t circ_buffer_max_size(circular_buffer_t *buff)
+{
     return buff->max_size;
 }
 
 /* Add single char to circular buffer */
-void circ_buff_push_char(circular_buffer_t *buff, const char c) {
+void circ_buff_push_char(circular_buffer_t *buff, const char c)
+{
     *(buff->array + buff->tail) = c;
     buff->tail = (buff->tail + 1) % buff->max_size;
 
@@ -68,9 +72,10 @@ void circ_buff_push_char(circular_buffer_t *buff, const char c) {
     }
 }
 
-/* Add string to circular buffer 
+/* Add string to circular buffer
  * If string size is too big, only the first chars will be copied */
-void circ_buff_push_string(circular_buffer_t *buff, const char *str, size_t str_size) {
+void circ_buff_push_string(circular_buffer_t *buff, const char *str, size_t str_size)
+{
     if (str_size >= buff->max_size) {
         memcpy(buff->array, str, buff->max_size);
         buff->head = 0;
@@ -85,10 +90,11 @@ void circ_buff_push_string(circular_buffer_t *buff, const char *str, size_t str_
 }
 
 /* Read last char in circular buffer */
-char circ_buff_read_char(circular_buffer_t *buff, int pos) {
+char circ_buff_read_char(circular_buffer_t *buff, int pos)
+{
     long cursor;
 
-    if (pos >= 0){
+    if (pos >= 0) {
         /* count up */
         cursor = (buff->head + pos) % buff->max_size;
     } else {
@@ -100,7 +106,8 @@ char circ_buff_read_char(circular_buffer_t *buff, int pos) {
 }
 
 /* Read certain size of the circular buffer */
-void circ_buff_read_all(circular_buffer_t *buff, char *buffer_out, size_t read_size) {
+void circ_buff_read_all(circular_buffer_t *buff, char *buffer_out, size_t read_size)
+{
     size_t len;
     if (buff->tail > buff->head) {
         len = MIN(read_size, (buff->tail - buff->head));
@@ -110,25 +117,26 @@ void circ_buff_read_all(circular_buffer_t *buff, char *buffer_out, size_t read_s
         len = MIN(read_size, top);
         memcpy(buffer_out, buff->array + buff->head, len);
 
-        len = MIN((read_size-top), buff->tail);
-        memcpy(buffer_out+top, buff->array, len);
+        len = MIN((read_size - top), buff->tail);
+        memcpy(buffer_out + top, buff->array, len);
     }
 }
 
 /* Clear circular buffer */
-void circ_buff_remove_n_chars(circular_buffer_t *buff, int pos) {
-    if (circ_buff_current_size(buff)>= abs(pos)) {
+void circ_buff_remove_n_chars(circular_buffer_t *buff, int pos)
+{
+    if (circ_buff_current_size(buff) >= (unsigned)abs(pos)) {
         if (pos < 0) {
             buff->tail = (buff->tail + pos) % buff->max_size;
-        }
-        else {
+        } else {
             buff->head = (buff->head + pos) % buff->max_size;
         }
     }
 }
 
 /* Clear circular buffer */
-void circ_buff_flush(circular_buffer_t *buff) {
+void circ_buff_flush(circular_buffer_t *buff)
+{
     buff->full = false;
     buff->head = buff->tail;
 }
